% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_clustering.R
\name{adj.rand.index}
\alias{adj.rand.index}
\title{Compute adjusted Rand Index between two clusters to evalute cluster similarity.}
\usage{
adj.rand.index(c1, c2)
}
\arguments{
\item{c1}{A vector containing cluster labels for a clustering}

\item{c2}{A vector containing cluster labels for another clustering}
}
\value{
a numeric between 0 and 1.
}
\description{
Compute adjusted Rand Index between two clusters to evalute cluster similarity.
}
