% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_clustering.R
\name{assign_cluster_celltypes}
\alias{assign_cluster_celltypes}
\title{Automatically assign cell type based on a Z test.}
\usage{
assign_cluster_celltypes(
  cL_,
  p.adj.cutoff = 0.1,
  fold.diff = 3,
  p.adj.range = 0.1
)
}
\arguments{
\item{cL_}{A matrix. A matrix of cell-type scores with clusters in the row and
cell types in the column. Matrix should be row-standardized (z scores). Output
of \code{sctype_score} function in Ianevski et al. (2021). Nat Commun.}

\item{p.adj.cutoff}{A numeric. Threshold for detecting a single cell type for
a cluster.}

\item{fold.diff}{A numeric. The factor by which the second smallest adjusted
p value must be greater than the smallest adjusted p value, for the cell type
with the smallest adjusted p value to be used to annotate the cluster.}

\item{p.adj.range}{A numeric. Use cell types with adjusted p value within this
range beyond the smallest adjusted p to annotate the cluster.}
}
\value{
a vector of characters specifying the assigned cell type for each cluster
}
\description{
Automatically assign cell type based on a Z test.
}
