% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t4_DE_test_glmm.R
\name{check_args_glmm}
\alias{check_args_glmm}
\title{Check validity of arguments for DE_glmm function.}
\usage{
check_args_glmm(
  sce,
  group_colname,
  random_var_colname,
  normalization_method,
  test_method,
  extra_params = list()
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{group_colname}{A string specifying the grouping column name}

\item{random_var_colname}{A string specifying the random variable column name}

\item{normalization_method}{A string specifying the normalization method}

\item{test_method}{A string specifying the test method}

\item{extra_params}{A list of additional parameters passed via \code{...}}
}
\value{
Logical.
}
\description{
Check validity of arguments for DE_glmm function.
}
\keyword{internal}
