% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_DE_test_pseudobulk.R
\name{check_args_pseudobulk}
\alias{check_args_pseudobulk}
\title{Check validity of arguments for DE_pseudobulk function}
\usage{
check_args_pseudobulk(
  sce,
  group_colname,
  normalization_method,
  test_method,
  extra_params = list()
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{group_colname}{A string specifying the grouping column name}

\item{normalization_method}{A string specifying the normalization method}

\item{test_method}{A string specifying the test method}

\item{extra_params}{A list of additional parameters passed via \code{...}}
}
\value{
Logical.
}
\description{
Check validity of arguments for DE_pseudobulk function
}
\keyword{internal}
