% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{cluster_components}
\alias{cluster_components}
\title{Cluster NMF components using optimized parameters.}
\usage{
cluster_components(nmf_w_list_filtered, snn_k, snn_type, leiden_res)
}
\arguments{
\item{nmf_w_list_filtered}{A list of filtered NMF weights.}

\item{snn_k}{An integer, specifying the number of nearest neighbors to
consider during graph construction.}

\item{snn_type}{A string, specifying the type of weighting scheme to use
for shared neighbors.}

\item{leiden_res}{A numeric, specifying the leiden resolution parameter to use.}
}
\value{
A list of length 4, including the leiden clustering outputs (\code{clust_leiden}),
the observed and expected edge weights (\code{edge_weight_leiden}), the modularity
(\code{mod_leiden}), and silhouette width (\code{sil.approx}).
}
\description{
Cluster NMF components using optimized parameters.
}
