% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_clustering.R
\name{cluster_one_fold}
\alias{cluster_one_fold}
\title{Cluster cells at multiple leiden resolutions, and returns an evaluation of the clustering
by computing a \code{deScore}.}
\usage{
cluster_one_fold(
  expr_mat,
  leiden_resolution_list,
  init_pca_k = 50,
  pca_var_cutoff = 0.8,
  snn_type = c("rank", "number", "jaccard"),
  snn_k = 10,
  used_cells = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{expr_mat}{gene-by-cell expression matrix, log2 normalized counts ("logcounts").}

\item{leiden_resolution_list}{numeric vector, a list of resolutions for leiden clustering.}

\item{init_pca_k}{A numeric, initial PCA dimension. Default is 50.}

\item{pca_var_cutoff}{A numeric, maximum variance explained in retained PCs. Default is 0.80.}

\item{snn_type}{A string, the type of weighting scheme for shared neighbors in kNN graphs.}

\item{snn_k}{An integer, the number of nearest neighbors to consider during graph construction.}

\item{used_cells}{A vector, list of indices for cells to use. If NULL, use all cells.}

\item{verbose}{A logical, whether to print progress messages. Default is FALSE.}

\item{...}{Additional arguments passed to \code{compute_hvg} (e.g., \code{fdr_top_cutoff}, \code{n_top_genes})
and \code{compute_deScores} (e.g., \code{min_cluster_size}, \code{min_max_de_score}, \code{min_mean_de_score}).}
}
\value{
A list of clustering outputs.
}
\description{
Cluster cells at multiple leiden resolutions, and returns an evaluation of the clustering
by computing a \code{deScore}.
}
