% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{cluster_opt}
\alias{cluster_opt}
\title{Optimizing clustering NMF components via \code{leiden} clustering}
\usage{
cluster_opt(
  nmf_w_list_filtered,
  snn_k_list = c(5, 10, 15, 20, 25),
  snn_type_list = c("rank", "number", "jaccard"),
  leiden_resolution_list = c(0.4, 0.7, 1)
)
}
\arguments{
\item{nmf_w_list_filtered}{A list of fitlered NMF components}

\item{snn_k_list}{A list of integers, specifying the number of nearest
neighbors to consider during SNN graph construction.}

\item{snn_type_list}{A list of string, specifying the type of weighting
scheme to use for shared neighbors.}

\item{leiden_resolution_list}{A list of numeric, specifying resolutions of
leiden clustering.}
}
\value{
A list of length 4, including the optimization outputs (\code{clust_leiden_opt}),
the silhouette width (\code{sil_df}), mean silhouette width (\code{sil_df_mean}), and a
plot of the mean silhouette width (\code{mean_silhouette_plot}).
}
\description{
Optimizing clustering NMF components via \code{leiden} clustering
}
