% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_clustering.R
\name{compute_deScores}
\alias{compute_deScores}
\title{Compute the \code{deScore} between clusters for a list of cluster memberships using pairwise
Wilcoxon tests.}
\usage{
compute_deScores(
  membership,
  expr_mat,
  min_cluster_size = 50,
  min_max_de_score = 5,
  min_mean_de_score = 1
)
}
\arguments{
\item{membership}{A vector specifying cluster memberships.}

\item{expr_mat}{A matrix of gene expression, log2 normalized counts ("logcounts").}

\item{min_cluster_size}{An integer, minimum cluster size to be considered valid.}

\item{min_max_de_score}{A numeric, minimum maximum pairwise deScore for a cluster to be considered valid.}

\item{min_mean_de_score}{A numeric, minimum mean pairwise deScore for a cluster to be considered valid.}
}
\value{
A data.frame summarizing mean deScores for the cluster memberships.
}
\description{
Compute the \code{deScore} between clusters for a list of cluster memberships using pairwise
Wilcoxon tests.
}
