% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_basics.R
\name{compute_hvg}
\alias{compute_hvg}
\title{Compute highly-variable genes for input matrix with given criteria}
\usage{
compute_hvg(expr_mat, used_cells, fdr_top_cutoff = 0.5, n_top_genes = 3000)
}
\arguments{
\item{expr_mat}{dgCMatrix, expression matrix, normalized, log2-transformed}

\item{used_cells}{logical, used cells}

\item{fdr_top_cutoff}{numeric (0-1), FDR cutoff for top variable gene selection based}

\item{n_top_genes}{numeric, number of genes with top variability. Ignored if \code{fdr_top_cutoff} is not null}
}
\value{
A logical vector
}
\description{
Compute highly-variable genes for input matrix with given criteria
}
