% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_basics.R
\name{compute_z_score}
\alias{compute_z_score}
\title{Compute z score for a vector
Adapted from: https://github.com/AllenInstitute/scrattch.hicat/blob/master/R/vg.R#L122
This version of z scoring will differ from base R's scale()}
\usage{
compute_z_score(vec)
}
\arguments{
\item{vec}{a numeric vector}
}
\value{
a numeric vector of z-scores
}
\description{
Compute z score for a vector
Adapted from: https://github.com/AllenInstitute/scrattch.hicat/blob/master/R/vg.R#L122
This version of z scoring will differ from base R's scale()
}
