% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{consensus_nmf}
\alias{consensus_nmf}
\title{Compute consensus NMFs}
\usage{
consensus_nmf(
  nmf_w_list_filtered,
  sil.approx.filtered,
  cnmf_type = c("median", "prototype")
)
}
\arguments{
\item{nmf_w_list_filtered}{A list of filtered NMF weights.}

\item{sil.approx.filtered}{A data.frame of mean silhouette scores for valid
clusters. Output of \code{post_clustering_filter}}

\item{cnmf_type}{A string specifying the type of cNMF to compute. One of \code{median}
(using median of components in a cluster), or \code{prototype} (using component
closest to the cluster centroid).}
}
\value{
A matrix of consensus NMF weights, or NULL.
}
\description{
Compute consensus NMFs
}
