% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{consensus_nmf_usage}
\alias{consensus_nmf_usage}
\title{Compute consensus NMF usage}
\usage{
consensus_nmf_usage(expr_mat, consensus_nmf)
}
\arguments{
\item{expr_mat}{An expression matrix with genes in the row and cells in the column.}

\item{consensus_nmf}{A matrix of consensus NMF weights.}
}
\value{
A list of length 3, including the weights (\code{cnmf_w}), usage (\code{cnmf_h}), and
the diagonal (\code{cnmf_d}) of the consensus NMF.
}
\description{
Compute consensus NMF usage
}
