% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t4_DE_test_glmm.R
\name{DE_glmm}
\alias{DE_glmm}
\title{Perform DE test using GLMM methods}
\usage{
DE_glmm(
  sce,
  group_colname,
  random_var_colname,
  normalization_method = c("RLE", "TMM", "colSums"),
  test_method = c("NBGMM", "NBLMM-reml", "PMM"),
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object of the single-cell data}

\item{group_colname}{A string. The grouping variable}

\item{random_var_colname}{A string. Name of the column that specifies the random variable.}

\item{normalization_method}{A string. Normalization method to use. One of "RLE", "TMM", and "colSums".}

\item{test_method}{A string. Distribution model to use. One of "NBGMM", "NBLMM-reml", and "PMM".}

\item{...}{Additional parameters, including \code{cpc}, \code{mincp}, and \code{ncore}.}
}
\value{
A list.
}
\description{
Perform DE test using GLMM methods
}
