% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_DE_test_pseudobulk.R
\name{DE_pseudobulk}
\alias{DE_pseudobulk}
\title{Perform differential expression test using pseudobulking methods.}
\usage{
DE_pseudobulk(
  sce,
  group_colname,
  normalization_method = c("RLE", "TMM"),
  test_method = c("DESeq2_pseudobulk", "edgeR_pseudobulk", "limma_pseudobulk"),
  ...
)
}
\arguments{
\item{sce}{An SingleCellExperiment object of the data. \code{sce} should be prefiltered to contain only valid samples}

\item{group_colname}{A string specifying the column (a factor) in the colData of \code{sce} to be used as grouping vairable}

\item{normalization_method}{A string specifying the normalization to use. One of \code{RLE}, and \code{TMM}}

\item{...}{Additional arguments. Valid options include: \code{min.count}, \code{min.total.count},
\code{DESeq2.test}, \code{DESeq2.fitType}, \code{DESeq2.sfType}, \code{DESeq2.lfcShrink_type}, \code{edgeR.test}.}

\item{test_methood}{A string specifying the pseudobulk method to use.}
}
\value{
A list of outputs.
}
\description{
Perform differential expression test using pseudobulking methods.
}
