% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_DE_test_pseudobulk.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Generate diagnostic plots for given outputs from the \code{DE_pseudobulk} function}
\usage{
diagnostic_plots(outs, output_dir)
}
\arguments{
\item{outs}{outputs from the \code{DE_pseudobulk} function}

\item{output_dir}{A string. Output directory to save plots}
}
\value{
logical.
}
\description{
Generate diagnostic plots for given outputs from the \code{DE_pseudobulk} function
}
