% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_basics.R
\name{.FindVariableFeatures_vst}
\alias{.FindVariableFeatures_vst}
\title{Replementation of the Seurat's \code{FindVariableFeatures} to apply Variance
Stabilizing Transformation on a matrix or dgCMatrix object.}
\usage{
.FindVariableFeatures_vst(object, loess.span = 0.3, clip.max = "auto")
}
\arguments{
\item{loess.span}{A numeric. Loess span parameter for VST}

\item{clip.max}{A numeric or "auto". The maximum value allowed after standardization.}

\item{counts}{A Matrix or inheriting from Matrix.}
}
\value{
A data.frame, specifying the mean, variance, expected variance
based on the mean-variance trend, and the standardized expected variance for
each gene.
}
\description{
Replementation of the Seurat's \code{FindVariableFeatures} to apply Variance
Stabilizing Transformation on a matrix or dgCMatrix object.
}
\keyword{internal}
