% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_basics.R
\name{estimate_sf_sparse}
\alias{estimate_sf_sparse}
\title{Compute size factor for a counts matrix}
\usage{
estimate_sf_sparse(counts, round_exprs = TRUE)
}
\arguments{
\item{counts}{A Matrix or inheriting from Matrix.}

\item{round_exprs}{A logical.}
}
\value{
A numeric vector
}
\description{
Compute size factor for a counts matrix
}
