% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_pathway_analysis.R
\name{extract_list_from_env}
\alias{extract_list_from_env}
\title{Extracts pathways from an environment containing the pathway list.}
\usage{
extract_list_from_env(
  used_env,
  type = c("GO", "Reactome", "KEGG"),
  ensembl_version = "110"
)
}
\arguments{
\item{used_env}{An environment containing the pathway list.}

\item{type}{A string, specifying the type of pathway. One of "GO", "Reactome",
and "KEGG"}

\item{ensembl_version}{A string, specifying the Ensembl version to use.}
}
\value{
A list. List of pathways (names and Ensembl IDs).
}
\description{
Extracts pathways from an environment containing the pathway list.
}
