% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_pathway_analysis.R
\name{extract_SynGO_pathways}
\alias{extract_SynGO_pathways}
\title{Extract SynGO pathways}
\usage{
extract_SynGO_pathways(syngo_path, ensembl_version = "110")
}
\arguments{
\item{syngo_path}{path to SynGO downloads}
}
\value{
A list. List of SynGO pathways in gene symbol
}
\description{
Extract SynGO pathways
}
