% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{filter_cnmf}
\alias{filter_cnmf}
\title{Filter cNMFs that capture mainly random effects. Test random effect by Wilcoxon
tests.}
\usage{
filter_cnmf(nmf_model, grouping, fdr_cutoff = 0.1)
}
\arguments{
\item{nmf_model}{A list, the NMF model object, include \code{cnmf_w}, \code{cnmf_d} and \code{cnmf_h}.}

\item{grouping}{A vector. A vector of string or factor, specifying the random factor
for the examples (column of the \code{h} matrix).}

\item{fdr_cutoff}{A numeric. The cutoff applied to FDR to detect outlier.}
}
\value{
A list of data.frame. Each data.frame contains the Wilcoxon test statistics
for all cNMFs for a level in \code{grouping}.
}
\description{
Filter cNMFs that capture mainly random effects. Test random effect by Wilcoxon
tests.
}
