% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{filter_components}
\alias{filter_components}
\title{Apply threshold on mean distance to \code{L} nearest neighbors, and filter components.}
\usage{
filter_components(nmf_w_list, graph_mat_mean, density_threshold = 0.3)
}
\arguments{
\item{nmf_w_list}{A list of normalized NMF weights.}

\item{graph_mat_mean}{A matrix of mean distance to \code{L} nearest neighbors.}

\item{density_threshold}{A numeric. Threshold on mean distance to \code{L} NNs.}
}
\value{
A list of filtered NMF weights.
}
\description{
Apply threshold on mean distance to \code{L} nearest neighbors, and filter components.
}
