% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_DE_test_pseudobulk.R
\name{filter_samples_pseudobulk}
\alias{filter_samples_pseudobulk}
\title{Filter samples for differential expression test.}
\usage{
filter_samples_pseudobulk(
  sce,
  group_by_var,
  n_cells_dataset,
  n_datasets_celltype
)
}
\arguments{
\item{sce}{SingleCellExperiment object of the aggregated data, output of \code{scater::aggregateAcrossCells}.}

\item{group_by_var}{A string. The grouping variable.}

\item{n_cells_dataset}{A numeric. The minimum number of cells per dataset for
the celltype to be considered valid}

\item{n_datasets_celltype}{A numeric. The mininum number of replicates (samples)
per condition in the \code{group_by_var} variable}
}
\value{
A SingleCellExperiment object representing the data with valid samples
if the data is valid. Or NULL if the data is invalid.
}
\description{
Filter samples for differential expression test.
}
