% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_pathway_analysis.R
\name{gsea_dotplot}
\alias{gsea_dotplot}
\title{Plot enrichment analysis results}
\usage{
gsea_dotplot(
  df,
  title_str,
  p_value_str = "fdr",
  p_value_threshold = 0.05,
  n_limit = NULL,
  n_min = 5,
  y_order_by = "NES",
  color_by = "fdr",
  size_by = "gene_ratio",
  pathway_label = "pathway_name",
  pathway_type_label = "pathway_type",
  x_axis_limits = NULL,
  colorbar_limits = NULL
)
}
\arguments{
\item{df}{A data.frame. GSEA outputs}

\item{title_str}{A string. Plot title.}

\item{p_value_str}{A string. Column name in the data.frame to be used as the probability value.}

\item{p_value_threshold}{A numeric. Probability value threshold.}

\item{n_limit}{A numeric. Maxiumum number of pathways to plot. If \code{NULL}, plot
all pathways.}

\item{n_min}{A numeric. Minimum number of pathways to plot.}

\item{y_order_by}{A string. Variable name for x axis and ordering.}

\item{color_by}{A string. Variable name for color.}

\item{size_by}{A string. Variable name for dot size.}

\item{pathway_label}{A string. Variable name for pathway labels.}

\item{pathway_type_label}{A string. Variable name for pathway label annotations.}

\item{x_axis_limits}{A nuermic vector of size 2, specifying the x axis limits.}

\item{colorbar_limits}{A nuermic vector of size 2, specifying the color legend limits.}
}
\value{
A ggplot2 object or NULL
}
\description{
Plot enrichment analysis results
}
