% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{load_model_weights}
\alias{load_model_weights}
\title{Load model weights and perform L2 normalization}
\usage{
load_model_weights(k_used, nmf_model_dir, seed_list)
}
\arguments{
\item{k_used}{An integer. Rank used.}

\item{nmf_model_dir}{A string. Directory to store nmf models.}

\item{seed_list}{A vector of random seeds (numeric)}
}
\value{
A list of normalized NMF weights.
}
\description{
Load model weights and perform L2 normalization
}
