% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_basics.R
\name{make_summed_sce}
\alias{make_summed_sce}
\title{Make mock aggregated data}
\usage{
make_summed_sce(ncol_ = 500, nrow_ = 400)
}
\arguments{
\item{ncol}{numeric, number of columns}

\item{nrow}{numeric, number of rows}
}
\value{
A SingleCellExperiment object
}
\description{
Make mock aggregated data
}
