% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_pathway_analysis.R
\name{map_human_gene_to_mouse_gene_symbol}
\alias{map_human_gene_to_mouse_gene_symbol}
\title{Map human genes (Ensembl ID or gene symbol) to mouse gene symbol}
\usage{
map_human_gene_to_mouse_gene_symbol(
  gene_list,
  input_type = c("ensembl_id", "entrez_id", "gene_symbol"),
  ensembl_version = "110"
)
}
\arguments{
\item{gene_list}{A list of human genes}

\item{input_type}{A string. Type of the input human genes. One of \code{ensembl_id},
\code{entrez_id}, and \code{gene_symbol}}

\item{ensembl_version}{A string. Ensembl version used.}
}
\value{
A data.frame. The first columns is the input human gene, and the
second the associated mouse Ensembl ID
}
\description{
Map human genes (Ensembl ID or gene symbol) to mouse gene symbol
}
