% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{mean_nn_distance}
\alias{mean_nn_distance}
\title{Find average distance to \code{L} nearest neighbors, and filter components.}
\usage{
mean_nn_distance(
  nmf_w_list,
  used_k,
  local_neighborhood_size = 0.3,
  distance_histogram_threshold = 0.3
)
}
\arguments{
\item{nmf_w_list}{A list of normalized NMF weights.}

\item{used_k}{An integer. Rank used.}

\item{local_neighborhood_size}{A numeric, specifies the local neighborhood size,
and controls the number of nearest neighbors to consider.}

\item{distance_histogram_threshold}{A numeric. A vertical cutoff to plot in the
histogram of mean distance to \code{L} nearest neighbor.}
}
\value{
A list of length 2, including \code{graph_mat_mean} and \code{histogram}.
}
\description{
Find average distance to \code{L} nearest neighbors, and filter components.
}
