% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{plot_clustergram}
\alias{plot_clustergram}
\title{Plot clustergram before or after filtering.}
\usage{
plot_clustergram(nmf_w_list, used_k, plot_title)
}
\arguments{
\item{nmf_w_list}{A list of normalized NMF weights.}

\item{used_k}{An integer. Rank used.}

\item{plot_title}{A string. Plot title}
}
\value{
A pheatmap object.
}
\description{
Plot clustergram before or after filtering.
}
