% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t6_plotting.R
\name{plot_data_}
\alias{plot_data_}
\title{Internal function to compute marker data for plotting stacked violin or heatmap.}
\usage{
plot_data_(sce, marker_dict, row_ensembl = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{marker_dict}{A list of markers in a format of list("A"=c('gene1','gene2',...)).
as the grouping variable.}

\item{row_ensembl}{Logical. Column names of \code{sce} is Ensembl ID or not.}
}
\value{
A \code{dgCMatrix} matrix of expression, with each row being a sample (cell),
and each column a marker gene.
}
\description{
Internal function to compute marker data for plotting stacked violin or heatmap.
}
