% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t6_plotting.R
\name{plot_marker_heatmap}
\alias{plot_marker_heatmap}
\title{Plot heatmap of markers}
\usage{
plot_marker_heatmap(
  sce,
  marker_dict,
  cluster_lab,
  row_ensembl = TRUE,
  scale_expr = TRUE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  plot_title = "heatmap"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{marker_dict}{A list of markers in a format of list("A"=c('gene1','gene2',...)).}

\item{cluster_lab}{A string. Specifying a column in colData(sce) to be used
as the grouping variable.}

\item{row_ensembl}{Logical. Column names of \code{sce} is Ensembl ID or not.}

\item{scale_expr}{Logical. Scale expression for each gene or not.}

\item{cluster_rows}{Logical. Cluster heatmap rows (genes) or not.}

\item{cluster_cols}{Logical. Cluster heatmap columns (groups) or not.}

\item{plot_title}{A string. Plot title.}
}
\value{
A \code{pheatmap} object.
}
\description{
Plot heatmap of markers
}
