% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t6_plotting.R
\name{plot_reduced_dimension}
\alias{plot_reduced_dimension}
\title{Plot 2D scatter plot for a reduced dimension}
\usage{
plot_reduced_dimension(
  sce,
  color_by,
  color_by_label,
  plot_title,
  show_label = FALSE,
  reduced_dim_name = "UMAP",
  reduced_dim_1 = "V1",
  reduced_dim_2 = "V2",
  assay.type = "logcounts",
  split_by = NULL,
  alpha = 0.5,
  size = 0.1,
  rasterize = FALSE,
  rastr_dpi = 600
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{color_by}{A string specifying color of the data point. One of the
column variables of colData of sce (discrete or continuous), gene name,
or a numeric vector.}

\item{color_by_label}{A string. Labels for color.}

\item{plot_title}{A string. Title of plot.}

\item{show_label}{Logical. Show cluster labels over data point or not.}

\item{reduced_dim_name}{A string. Name of the reduced dimension to use.}

\item{reduced_dim_1}{A string. Column name of first reduced dimension.}

\item{reduced_dim_2}{A string. Column name of second reduced dimension.}

\item{assay.type}{A string. The type of \code{assay} to use when plotting expression
of a gene.}

\item{split_by}{A string. Name of a column in \code{colData} used to split the scatter
plot into subplots.}

\item{alpha}{A numeric. Alpha value for data points.}

\item{size}{A numeric. Size value for data points.}

\item{rasterize}{Logical. To rasterize the scatter plot or not.}

\item{rastr_dpi}{An integer. dpi value for rasterizing.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot 2D scatter plot for a reduced dimension
}
