% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t6_plotting.R
\name{plot_stackedViolin}
\alias{plot_stackedViolin}
\title{Plot stacked violin plots for genes}
\usage{
plot_stackedViolin(
  sce,
  marker_dict,
  cluster_lab,
  row_ensembl = TRUE,
  plot_title = "stacked violin plot",
  x_lab = "cluster",
  y_lab = "expression level (log)"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{marker_dict}{A list of markers in a format of list("A"=c('gene1','gene2',...)).}

\item{cluster_lab}{A string. Specifying a column in colData(sce) to be used
as the grouping variable.}

\item{row_ensembl}{Logical. Column names of \code{sce} is Ensembl ID or not.}

\item{plot_title}{A string. Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot stacked violin plots for genes
}
