% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{post_clustering_filter}
\alias{post_clustering_filter}
\title{Apply post-clustering filtering on NMF components.}
\usage{
post_clustering_filter(
  sil.approx,
  sil_approx_thresh = 0,
  cluster_size_thresh = 50
)
}
\arguments{
\item{sil.approx}{A data.frame of silhouette width.}

\item{sil_approx_thresh}{A numeric, cutoff for filtering clusters with silhouette
width lower than this value.}

\item{cluster_size_thresh}{An integer, minimum cluster size to keep.}
}
\value{
A data.frame of stats (mean silhouette) of filtered clusters.
}
\description{
Apply post-clustering filtering on NMF components.
}
