% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_pathway_analysis.R
\name{process_pathway}
\alias{process_pathway}
\title{Process a pathway (Gene Ontology, Reactome, KEGG, SynGO)}
\usage{
process_pathway(
  pathway_output_path,
  pathway_db_ver,
  pathway_str = c("GO_BP", "GO_CC", "GO_MF", "Reactome", "KEGG", "SynGO"),
  ensembl_version = "110",
  syngo_path = NULL
)
}
\arguments{
\item{pathway_output_path}{A String, output pathway}

\item{pathway_str}{A String, specifies type of pathways (one of GO_BP, GO_CC,
GO_MF, Reactome, KEGG, SynGO)}

\item{pathway_db_str}{A String, version of processed pathways}
}
\value{
TRUE if no error.
}
\description{
Process a pathway (Gene Ontology, Reactome, KEGG, SynGO)
}
