% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_consensus_nmf.R
\name{run_NMF_iter}
\alias{run_NMF_iter}
\title{Run NMF decomposition for multiple iterations.}
\usage{
run_NMF_iter(expr_mat, k_used, nmf_model_dir, seed_list)
}
\arguments{
\item{expr_mat}{A dense or sparse matrix of features in rows and samples in
columns. Prefer \code{matrix} or \code{Matrix::dgCMatrix}. Log2 normalized counts.}

\item{k_used}{An integer. Rank used.}

\item{nmf_model_dir}{A string. Directory to store nmf models.}

\item{seed_list}{A vector of random seeds numeric.}
}
\value{
Logical. TRUE if no error.

Logical. TRUE if no error.
}
\description{
Run NMF decomposition for multiple iterations.
}
